requirements.txt
Django==2.0.3
coverage==4.5.1
dj-database-url==0.5.0
django-allauth==0.35.0
django-environ==0.4.4
django-extensions==2.0.0
django-paypal==0.5.0
psycopg2==2.7.4
model-mommy==1.5.1
gunicorn==19.7.1
pycrypto==2.6.1
simple-crypt==4.1.7
django-braces==1.12.0
python-dateutil==2.6.1
python-memcached==1.59
django-debug-toolbar==1.9.1
django-crispy-forms==1.7.0
django-apptemplates==1.4
shortuuid==0.5.0
xlrd==1.1.0
xlwt==1.3.0